+++ api/webiny.application.ts
@@ -5,5 +5,12 @@
 export default {
     id: "api",
     name: "API",
-    description: "Your GraphQL API and all of the backend services."
+    description: "Your GraphQL API and all of the backend services.",
+    cli: {
+        // Default args for the "yarn webiny watch ..." command.
+        watch: {
+            // Watch five levels of dependencies, starting from this project application.
+            depth: 5
+        }
+    }
 };