+++ packages/app-headless-cms/src/admin/plugins/fieldRenderers/ref/components/renderItem.tsx
@@ -1,17 +1,25 @@
 import React from "react";
 import styled from "@emotion/styled";
 import { Typography } from "@webiny/ui/Typography";
+import { createEntryUrl } from "~/admin/plugins/fieldRenderers/ref/components/createEntryUrl";
+import { Link } from "@webiny/react-router";
 
 const ModelId = styled("span")({
     color: "var(--mdc-theme-text-secondary-on-background) !important"
 });
 
-export const renderItem = item => {
+export interface Props {
+    name: string;
+    modelName: string;
+    modelId: string;
+    id: string;
+}
+export const renderItem = (props: Props) => {
     return (
         <Typography use={"body2"}>
-            {item.name}
+            <Link to={createEntryUrl(props)}>{props.name}</Link>
             <br />
-            <ModelId>Model: {item.modelName}</ModelId>
+            <ModelId>Model: {props.modelName}</ModelId>
         </Typography>
     );
 };