+++ packages/app/src/plugins/TenantHeaderLinkPlugin.ts
@@ -23,6 +23,8 @@ export class TenantHeaderLinkPlugin extends ApolloLinkPlugin {
         }
 
         this.tenant = tenant;
+
+        console.log("Detected tenant", this.tenant);
     }
 
     createLink() {