+++ packages/app-admin/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.18.3](https://github.com/webiny/webiny-js/compare/v5.18.3-beta.0...v5.18.3) (2021-12-02)
+
+**Note:** Version bump only for package @webiny/app-admin
+
+
+
+
+
+## [5.18.3-beta.0](https://github.com/webiny/webiny-js/compare/v5.18.2...v5.18.3-beta.0) (2021-12-01)
+
+
+### Bug Fixes
+
+* read app config before env vars ([147923f](https://github.com/webiny/webiny-js/commit/147923f5aebb901556afe334e5c346f17a239ed0))
+
+
+
+
+
+## [5.18.2](https://github.com/webiny/webiny-js/compare/v5.18.2-beta.0...v5.18.2) (2021-11-29)
+
+**Note:** Version bump only for package @webiny/app-admin
+
+
+
+
+
+## [5.18.2-beta.0](https://github.com/webiny/webiny-js/compare/v5.18.1...v5.18.2-beta.0) (2021-11-29)
+
+**Note:** Version bump only for package @webiny/app-admin
+
+
+
+
+
 ## [5.18.1](https://github.com/webiny/webiny-js/compare/v5.18.1-beta.0...v5.18.1) (2021-11-25)
 
 **Note:** Version bump only for package @webiny/app-admin